+++ src/assos/assos.controller.ts
@@ -1,5 +1,5 @@
 import { Body, Controller, Delete, Get, Patch, Post, Put, Query } from '@nestjs/common';
-import { ApiCreatedResponse, ApiOkResponse, ApiOperation, ApiQuery, ApiTags } from '@nestjs/swagger';
+import { ApiCreatedResponse, ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { ApiAppErrorResponse, paginatedResponseDto } from '../app.dto';
 import { AssoMembershipRole } from './interfaces/membership-role.interface';
 import { AssoMembership } from './interfaces/membership.interface';
@@ -23,12 +23,7 @@ import AssosMemberCreateReqDto from './dto/req/assos-member-create.dto';
 import AssosMemberUpdateReqDto from './dto/req/assos-member-update.dto';
 import AssoMembershipResDto from './dto/res/assos-membership-res.dto';
 import UsersService from '../users/users.service';
-import AssosPostWeeklyReqDto from './dto/req/assos-post-weekly-req.dto';
-import WeeklyResDto from './dto/res/weekly-res.dto';
 import { ConfigModule } from '../config/config.module';
-import AssoGetWeeklyReqDto from './dto/req/asso-get-weekly-req.dto';
-import { AssoWeekly } from './interfaces/weekly.interface';
-import WeeklyInfoResDto from './dto/res/weekly-info-res.dto';
 
 @Controller('assos')
 @ApiTags('Assos')
@@ -221,122 +216,6 @@ export class AssosController {
       throw new AppException(ERROR_CODE.PARAM_TOO_HIGH, 'position');
     const updatedRoles = await this.assosService.updateAssoRole(role.id, asso.id, pick(body, 'name', 'position'));
     return { roles: updatedRoles.map(this.formatPartialAssoMembershipRole) };
-  }
-
-  @Get('/weekly/info')
-  @ApiOperation({ description: 'Returns information about weeklies.' })
-  @ApiOkResponse({ type: WeeklyInfoResDto })
-  getWeeklyInfo(): WeeklyInfoResDto {
-    return {
-      sendDay: this.config.WEEKLY_SEND_DAY,
-      sendHour: this.config.WEEKLY_SEND_HOUR,
-    };
-  }
-
-  @Get('/:assoId/weekly')
-  @ApiOperation({ description: 'Get weeklies from query parameter `from` to query parameter `to`.' })
-  @ApiQuery({ name: 'from', type: String })
-  @ApiQuery({ name: 'to', type: String, default: `\`from\` + env.PAGINATION_PAGE_SIZE days`, required: false })
-  @ApiOkResponse({ type: WeeklyResDto })
-  @ApiAppErrorResponse(ERROR_CODE.PARAM_DATE_MUST_BE_AFTER, '`to` must come after `from` (or be equal)')
-  @ApiAppErrorResponse(
-    ERROR_CODE.FORBIDDEN_ASSOS_PERMISSIONS,
-    'The user issuing the request does not have the permission weekly',
-  )
-  async searchWeeklies(
-    @ParamAsso() asso: Asso,
-    @Query() { from, to, page }: AssoGetWeeklyReqDto,
-    @GetUser() user: User,
-  ): Promise<Pagination<WeeklyResDto>> {
-    from = from.dropTime();
-    to = to?.dropTime();
-    if (to && from > to)
-      throw new AppException(ERROR_CODE.PARAM_DATE_MUST_BE_AFTER, to.toISOString(), from.toISOString());
-    if (!(await this.assosService.hasSomeAssoPermission(asso, user.id, 'weekly')))
-      throw new AppException(ERROR_CODE.FORBIDDEN_ASSOS_PERMISSIONS, asso.id, 'weekly');
-    const { weeklies, count } = await this.assosService.searchWeeklies(asso.id, from, to, page);
-    return {
-      items: weeklies,
-      itemCount: count,
-      itemsPerPage: this.config.PAGINATION_PAGE_SIZE,
-    };
-  }
-
-  @Post('/:assoId/weekly')
-  @ApiOperation({ description: 'Create a weekly for the given association.' })
-  @ApiCreatedResponse({ type: WeeklyResDto })
-  @ApiAppErrorResponse(
-    ERROR_CODE.FORBIDDEN_ASSOS_PERMISSIONS,
-    'The user issuing the request does not have the permission weekly',
-  )
-  @ApiAppErrorResponse(ERROR_CODE.WEEKLY_ALREADY_SENT_FOR_WEEK, 'The weekly was already sent for the specified week')
-  @ApiAppErrorResponse(ERROR_CODE.WEEKLY_ALREADY_PLANNED_FOR_WEEK, 'The asso already has a weekly planned for the requested week')
-  async createWeekly(
-    @ParamAsso() asso: Asso,
-    @Body() dto: AssosPostWeeklyReqDto,
-    @GetUser() user: User,
-  ): Promise<WeeklyResDto> {
-    if (!(await this.assosService.hasSomeAssoPermission(asso, user.id, 'weekly')))
-      throw new AppException(ERROR_CODE.FORBIDDEN_ASSOS_PERMISSIONS, asso.id, 'weekly');
-    if (this.assosService.getWeeklySendDate(dto.date) < new Date())
-      throw new AppException(ERROR_CODE.WEEKLY_ALREADY_SENT_FOR_WEEK, dto.date.toISOString());
-    if (await this.assosService.hasWeekly(asso.id, dto.date))
-      throw new AppException(ERROR_CODE.WEEKLY_ALREADY_PLANNED_FOR_WEEK);
-    return this.assosService.addWeekly(asso.id, dto.title, dto.message, dto.date);
-  }
-
-  @Patch('/:assoId/weekly/:weeklyId')
-  @ApiOperation({ description: 'Update a weekly for the given association.' })
-  @ApiOkResponse({ type: paginatedResponseDto(WeeklyResDto) })
-  @ApiAppErrorResponse(ERROR_CODE.NO_SUCH_ASSO)
-  @ApiAppErrorResponse(
-    ERROR_CODE.FORBIDDEN_ASSOS_PERMISSIONS,
-    'The user issuing the request does not have the permission weekly',
-  )
-  @ApiAppErrorResponse(ERROR_CODE.NO_SUCH_WEEKLY, 'The weekly does not exist for the specified asso')
-  @ApiAppErrorResponse(ERROR_CODE.WEEKLY_ALREADY_SENT, 'The weekly that is beeing modified was already sent')
-  @ApiAppErrorResponse(ERROR_CODE.WEEKLY_ALREADY_SENT_FOR_WEEK, 'The weekly was already sent for the specified week')
-  @ApiAppErrorResponse(ERROR_CODE.WEEKLY_ALREADY_PLANNED_FOR_WEEK, 'The asso already has a weekly planned for the requested week')
-  async updateWeekly(
-    @ParamAsso() asso: Asso,
-    @UUIDParam('weeklyId') weeklyId: string,
-    @Body() dto: AssosPostWeeklyReqDto,
-    @GetUser() user: User,
-  ): Promise<WeeklyResDto> {
-    if (!(await this.assosService.hasSomeAssoPermission(asso, user.id, 'weekly')))
-      throw new AppException(ERROR_CODE.FORBIDDEN_ASSOS_PERMISSIONS, asso.id, 'weekly');
-    const weekly: AssoWeekly = await this.assosService.getWeekly(weeklyId, asso.id);
-    if (!weekly) throw new AppException(ERROR_CODE.NO_SUCH_WEEKLY, weeklyId);
-    if (this.assosService.getWeeklySendDate(weekly.date) < new Date())
-      throw new AppException(ERROR_CODE.WEEKLY_ALREADY_SENT);
-    if (this.assosService.getWeeklySendDate(dto.date) < new Date())
-      throw new AppException(ERROR_CODE.WEEKLY_ALREADY_SENT_FOR_WEEK, dto.date.toISOString());
-    if (await this.assosService.hasWeekly(asso.id, weekly.date, weekly.id))
-      throw new AppException(ERROR_CODE.WEEKLY_ALREADY_PLANNED_FOR_WEEK);
-    return this.assosService.updateWeekly(weeklyId, dto);
-  }
-
-  @Delete('/:assoId/weekly/:weeklyId')
-  @ApiOperation({ description: 'Delete a weekly for the given association.' })
-  @ApiOkResponse({ type: WeeklyResDto })
-  @ApiAppErrorResponse(
-    ERROR_CODE.FORBIDDEN_ASSOS_PERMISSIONS,
-    'The user issuing the request does not have the permission weekly',
-  )
-  @ApiAppErrorResponse(ERROR_CODE.NO_SUCH_WEEKLY)
-  @ApiAppErrorResponse(ERROR_CODE.WEEKLY_ALREADY_SENT, 'The weekly that is beeing modified was already sent')
-  async deleteWeekly(
-    @ParamAsso() asso: Asso,
-    @UUIDParam('weeklyId') weeklyId: string,
-    @GetUser() user: User,
-  ): Promise<WeeklyResDto> {
-    if (!(await this.assosService.hasSomeAssoPermission(asso, user.id, 'weekly')))
-      throw new AppException(ERROR_CODE.FORBIDDEN_ASSOS_PERMISSIONS, asso.id, 'weekly');
-    const weekly: AssoWeekly = await this.assosService.getWeekly(weeklyId, asso.id);
-    if (!weekly) throw new AppException(ERROR_CODE.NO_SUCH_WEEKLY, weeklyId);
-    if (this.assosService.getWeeklySendDate(weekly.date) < new Date())
-      throw new AppException(ERROR_CODE.WEEKLY_ALREADY_SENT);
-    return this.assosService.deleteWeekly(weeklyId);
   }
 
   formatAssoOverview(asso: Asso): AssoOverviewResDto {
