+++ src/assos/assos.service.ts
@@ -2,15 +2,13 @@ import { Injectable } from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { ConfigModule } from '../config/config.module';
 import { PrismaService } from '../prisma/prisma.service';
-import { RawAssoMembershipRole, Translation } from '../prisma/types';
+import { RawAssoMembershipRole } from '../prisma/types';
 import { Asso } from './interfaces/asso.interface';
 import { AssoMembership } from './interfaces/membership.interface';
 import { AssoMembershipRole } from './interfaces/membership-role.interface';
 import AssosSearchReqDto from './dto/req/assos-search-req.dto';
 import AssosMemberUpdateReqDto from './dto/req/assos-member-update.dto';
 import { AppException, ERROR_CODE } from '../exceptions';
-import { AssoWeekly } from './interfaces/weekly.interface';
-import WeeklyResDto from './dto/res/weekly-res.dto';
 
 @Injectable()
 export class AssosService {
@@ -286,68 +284,5 @@ export class AssosService {
           : {}),
       },
     });
-  }
-
-  async searchWeeklies(assoId: string, from: Date, to: Date, page: number): Promise<{ weeklies: AssoWeekly[], count: number }> {
-    const where = {
-      assoId,
-      date: { gte: from, lte: to },
-    } satisfies Prisma.AssoWeeklyWhereInput;
-    const count = await this.prisma.assoWeekly.count({ where });
-    const weeklies = await this.prisma.normalize.assoWeekly.findMany({ where, skip: (page - 1) * this.config.PAGINATION_PAGE_SIZE, take: this.config.PAGINATION_PAGE_SIZE });
-    return { weeklies, count };
-  }
-
-  async addWeekly(assoId: string, title: Translation, message: Translation, date): Promise<AssoWeekly> {
-    return this.prisma.normalize.assoWeekly.create({
-      data: {
-        asso: { connect: { id: assoId } },
-        titleTranslation: { create: title },
-        bodyTranslation: { create: message },
-        date,
-      }
-    });
-  }
-
-  getWeeklySendDate(sendWeek: Date): Date {
-    return new Date(
-      Date.UTC(
-        sendWeek.getUTCFullYear(),
-        sendWeek.getUTCMonth(),
-        sendWeek.getUTCDate() + this.config.WEEKLY_SEND_DAY,
-        this.config.WEEKLY_SEND_HOUR,
-        0,
-        -Date.getTimezoneOffset('Europe/Paris')
-      )
-    );
-  }
-
-  async getWeekly(weeklyId: string, assoId?: string): Promise<WeeklyResDto> {
-    return this.prisma.normalize.assoWeekly.findUnique({ where: { id: weeklyId, assoId } });
-  }
-
-  async hasWeekly(assoId: string, date: Date, exclude: string = undefined): Promise<boolean> {
-    return (await this.prisma.assoWeekly.count({
-      where: {
-        assoId,
-        date,
-        ...(exclude ? { id: { not: exclude } } : {})
-      }
-    })) > 0;
-  }
-
-  async updateWeekly(weeklyId: string, fields: { title: Translation, message: Translation, date: Date }): Promise<AssoWeekly> {
-    return this.prisma.normalize.assoWeekly.update({
-      where: { id: weeklyId },
-      data: {
-        titleTranslation: { update: fields.title },
-        bodyTranslation: { update: fields.message },
-        date: fields.date,
-      }
-    });
-  }
-
-  async deleteWeekly(weeklyId: string): Promise<AssoWeekly> {
-    return this.prisma.normalize.assoWeekly.delete({ where: { id: weeklyId } });
   }
 }
