+++ src/assos/weekly/dto/res/weekly-info-res.dto.ts
